requirements.txt
matplotlib
keras
numpy
Pandas3
seaborn
tensorflow
scikit-learn